requirements_runtime.txt
Django==6.0.3
Pillow==12.2.0
openpyxl==3.1.5
reportlab==4.4.3
pywebview==5.4
